requirements.txt
Flask==2.0.1
Werkzeug==2.0.1
Pillow==11.0.0
flask-cors==4.0.0
